Update CrawlEventAdapter to move validation responsibility inside event adapter
Update PrimaryAdapter to set depth to number rather than string
