+++ services/crawl/functions/crawl-urls/adapters/__tests__/CrawlEventAdapter.test.ts
@@ -1,26 +1,24 @@
-import { ObjectValidator } from "@ashley-evans/buzzword-object-validator";
 import { mock } from "jest-mock-extended";
 
 import { CrawlPort } from "../../ports/CrawlPort";
 import { CrawlEvent, CrawlResponse } from "../../ports/PrimaryAdapter";
-import { CrawlEventAdapter, ValidCrawlEvent } from "../CrawlEventAdapter";
+import { CrawlEventAdapter } from "../CrawlEventAdapter";
 
 const mockCrawlPort = mock<CrawlPort>();
-const mockValidator = mock<ObjectValidator<ValidCrawlEvent>>();
 
-const adapter = new CrawlEventAdapter(mockCrawlPort, mockValidator);
+const adapter = new CrawlEventAdapter(mockCrawlPort);
 
 const EXPECTED_VALID_URL = new URL("http://www.example.com");
 const EXPECTED_PATHNAMES = ["/", "/test"];
 
-function createEvent(url?: URL | string, depth?: number | string): CrawlEvent {
+function createEvent(url?: URL | string, depth?: number): CrawlEvent {
     const event: CrawlEvent = {};
     if (url) {
         event.url = url.toString();
     }
 
     if (depth) {
-        event.depth = depth.toString();
+        event.depth = depth;
     }
 
     return event;
@@ -40,13 +38,7 @@ describe.each([
     beforeAll(async () => {
         jest.resetAllMocks();
 
-        mockValidator.validate.mockReturnValue({ url: eventURL });
         response = await adapter.crawl(event);
-    });
-
-    test("calls validator to validate event", () => {
-        expect(mockValidator.validate).toHaveBeenCalledTimes(1);
-        expect(mockValidator.validate).toHaveBeenCalledWith(event);
     });
 
     test("does not call crawl port", async () => {
@@ -56,39 +48,6 @@ describe.each([
     test("returns provided URL in response", () => {
         expect(response).toBeDefined();
         expect(response.baseURL).toEqual(eventURL);
-    });
-
-    test("returns failure", () => {
-        expect(response).toBeDefined();
-        expect(response.success).toEqual(false);
-    });
-
-    test("returns no pathnames array", () => {
-        expect(response).toBeDefined();
-        expect(response.pathnames).toBeUndefined();
-    });
-});
-
-describe("handles invalid event that does not pass validation", () => {
-    const event = createEvent();
-    let response: CrawlResponse;
-
-    beforeAll(async () => {
-        jest.resetAllMocks();
-
-        mockValidator.validate.mockImplementation(() => {
-            throw new Error();
-        });
-        response = await adapter.crawl(event);
-    });
-
-    test("calls validator with provided event", () => {
-        expect(mockValidator.validate).toHaveBeenCalledTimes(1);
-        expect(mockValidator.validate).toHaveBeenCalledWith(event);
-    });
-
-    test("does not call crawl port", async () => {
-        expect(mockCrawlPort.crawl).toHaveBeenCalledTimes(0);
     });
 
     test("returns failure", () => {
@@ -114,21 +73,12 @@ describe.each([
         beforeAll(async () => {
             jest.resetAllMocks();
 
-            mockValidator.validate.mockReturnValue({
-                url: expectedURL.toString(),
-                depth: expectedDepth,
-            });
             mockCrawlPort.crawl.mockResolvedValue({
                 success: true,
                 pathnames: EXPECTED_PATHNAMES,
             });
 
             response = await adapter.crawl(event);
-        });
-
-        test("calls object validator with provided event", () => {
-            expect(mockValidator.validate).toHaveBeenCalledTimes(1);
-            expect(mockValidator.validate).toHaveBeenCalledWith(event);
         });
 
         test("calls crawl port with URL and depth from event", () => {
@@ -161,10 +111,6 @@ describe("error handling", () => {
 
     beforeEach(() => {
         jest.resetAllMocks();
-
-        mockValidator.validate.mockReturnValue({
-            url: EXPECTED_VALID_URL.toString(),
-        });
     });
 
     test("throws crawl error if error occurs during crawl", async () => {
