+++ services/crawl/functions/crawl-urls/crawl-urls.ts
@@ -1,4 +1,3 @@
-import { JSONSchemaType } from "ajv";
 import {
     ContentRepository,
     S3Repository,
@@ -7,16 +6,9 @@ import {
     Repository,
     URLsTableRepository,
 } from "buzzword-aws-crawl-urls-repository-library";
-import {
-    ObjectValidator,
-    AjvValidator,
-} from "@ashley-evans/buzzword-object-validator";
 
 import { ApifyProvider } from "./adapters/ApifyProvider";
-import {
-    CrawlEventAdapter,
-    ValidCrawlEvent,
-} from "./adapters/CrawlEventAdapter";
+import { CrawlEventAdapter } from "./adapters/CrawlEventAdapter";
 import Crawl from "./domain/Crawl";
 import { CrawlProvider } from "./ports/CrawlProvider";
 import { CrawlEvent, CrawlResponse } from "./ports/PrimaryAdapter";
@@ -63,24 +55,6 @@ function createContentRepository(): ContentRepository {
     return new S3Repository(process.env.CONTENT_BUCKET_NAME);
 }
 
-function createValidator(): ObjectValidator<ValidCrawlEvent> {
-    const schema: JSONSchemaType<ValidCrawlEvent> = {
-        type: "object",
-        properties: {
-            url: {
-                type: "string",
-            },
-            depth: {
-                type: "integer",
-                nullable: true,
-            },
-        },
-        required: ["url"],
-    };
-
-    return new AjvValidator<ValidCrawlEvent>(schema, { coerceTypes: true });
-}
-
 const handler = async (event: CrawlEvent): Promise<CrawlResponse> => {
     const crawlProvider = createCrawlProvider();
     const urlRepository = createRepostiory();
@@ -92,9 +66,7 @@ const handler = async (event: CrawlEvent): Promise<CrawlResponse> => {
         contentRepository
     );
 
-    const validator = createValidator();
-
-    const primaryAdapter = new CrawlEventAdapter(crawlDomain, validator);
+    const primaryAdapter = new CrawlEventAdapter(crawlDomain);
 
     return await primaryAdapter.crawl(event);
 };
